trading/requirements.txt
# Trading System Monitoring - Python Dependencies

# Alpaca Trading API
alpaca-py>=0.8.0

# PostgreSQL Database
psycopg2-binary>=2.9.0

# Configuration
python-dotenv>=1.0.0

# Scheduling
APScheduler>=3.10.0
pytz>=2023.3

# Development/Testing Only
testing.postgresql>=1.3.0  # For in-memory PostgreSQL during testing
pytest>=7.4.0
pytest-cov>=4.1.0
